requirements.txt
langchain
langchain_core
langchain-anthropic
langsmith
langchain_voyageai
langchain_text_splitters